Staying up to date before cooldown with thermometry calibration

Merge branch 'master' into dhl88_bluefors_saa

+++ Utilities/save.py
@@ -247,34 +247,37 @@ class Measurement:
             nopdf = False
 
         ## Save remotely
-        try:
-            # Make sure directories exist
-            remote_dir = os.path.split(remote_path)[0]
-            if not os.path.exists(remote_dir):
-                os.makedirs(remote_dir)
+        if os.path.exists(get_data_server_path()):
+            try:
+                # Make sure directories exist
+                remote_dir = os.path.split(remote_path)[0]
+                if not os.path.exists(remote_dir):
+                    os.makedirs(remote_dir)
 
-            ## Loop over filetypes
-            for ext in ['.h5','.json','.pdf']:
-                if ext == '.pdf' and nopdf:
-                    continue
-                # First make a checksum
-                local_checksum = _md5(local_path + ext)
+                ## Loop over filetypes
+                for ext in ['.h5','.json','.pdf']:
+                    if ext == '.pdf' and nopdf:
+                        continue
+                    # First make a checksum
+                    local_checksum = _md5(local_path + ext)
 
-                # Copy the files
-                shutil.copyfile(local_path + ext, remote_path + ext)
+                    # Copy the files
+                    shutil.copyfile(local_path + ext, remote_path + ext)
 
-                # Make comparison checksums
-                remote_checksum = _md5(remote_path + ext)
+                    # Make comparison checksums
+                    remote_checksum = _md5(remote_path + ext)
 
-                # Check checksums
-                if local_checksum != remote_checksum:
-                    print('%s checksum failed! Cannot trust remote file %s' %(ext, remote_path + ext))
+                    # Check checksums
+                    if local_checksum != remote_checksum:
+                        print('%s checksum failed! Cannot trust remote file %s' %(ext, remote_path + ext))
 
-        except Exception as e:
-            if not os.path.exists(get_data_server_path()):
-                print('SAMBASHARE not connected. Could not find path %s. Object saved locally but not remotely.' %get_data_server_path())
-            else:
-                print('Saving to data server failed!\n\nException details: %s\n\nremote path: %s\nlocal path:%s' %(e, remote_path, local_path))
+            except Exception as e:
+                if not os.path.exists(get_data_server_path()):
+                    print('SAMBASHARE not connected. Could not find path %s. Object saved locally but not remotely.' %get_data_server_path())
+                else:
+                    print('Saving to data server failed!\n\nException details: %s\n\nremote path: %s\nlocal path:%s' %(e, remote_path, local_path))
+        else:
+            print('Not connected to %s, data not saved remotely!' %get_data_server_path())
 
         ## See if saving worked properly
         try:
@@ -415,11 +418,16 @@ def get_todays_data_dir():
     todays_data_path = os.path.join(experiment_path, now.strftime('%Y-%m-%d'))
 
     # Make local and remote directory
-    dirs = [get_local_data_path(), get_remote_data_path()]
+    dirs = [get_local_data_path()]
+    if os.path.exists(get_data_server_path()):
+        dirs += [get_remote_data_path()]
     for d in dirs:
-        filename = os.path.join(d, todays_data_path)
-        if not os.path.exists(filename):
-            os.makedirs(filename)
+        try:
+            filename = os.path.join(d, todays_data_path)
+            if not os.path.exists(filename):
+                os.makedirs(filename)
+        except Exception as e:
+            print(e)
 
     return todays_data_path
 
@@ -434,7 +442,9 @@ def set_experiment_data_dir(description=''):
     now_fmt = now.strftime('%Y-%m-%d')
 
     # Make local and remote directories:
-    dirs = [get_local_data_path(), get_remote_data_path()]
+    dirs = [get_local_data_path()]
+    if os.path.exists(get_data_server_path()):
+        dirs += [get_remote_data_path()]
     for d in dirs:
         try:
             filename = os.path.join(d, now_fmt + '_' + description)
